+++ packages/pytest-simcore/src/pytest_simcore/file_extra.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 from faker import Faker
 from pydantic import ByteSize
-from pytest_simcore.helpers.logging_ext import log_context
+from pytest_simcore.helpers.logging_tools import log_context
 
 
 @pytest.fixture
