+++ packages/pytest-simcore/src/pytest_simcore/helpers/playwright.py
@@ -9,7 +9,7 @@ from enum import Enum, unique
 from typing import Any, Final
 
 from playwright.sync_api import FrameLocator, Page, Request, WebSocket, expect
-from pytest_simcore.helpers.logging_ext import log_context
+from pytest_simcore.helpers.logging_tools import log_context
 
 SECOND: Final[int] = 1000
 MINUTE: Final[int] = 60 * SECOND
