+++ tests/e2e-playwright/tests/conftest.py
@@ -22,7 +22,7 @@ from playwright.sync_api import APIRequestContext, BrowserContext, Page, WebSock
 from playwright.sync_api._generated import Playwright
 from pydantic import AnyUrl, TypeAdapter
 from pytest import Item
-from pytest_simcore.helpers.logging_ext import log_context
+from pytest_simcore.helpers.logging_tools import log_context
 from pytest_simcore.helpers.playwright import (
     MINUTE,
     AutoRegisteredUser,
