+++ tests/e2e-playwright/tests/resource_usage_tracker/test_resource_usage_tracker.py
@@ -14,7 +14,7 @@ from http import HTTPStatus
 import pytest
 from playwright.sync_api import APIRequestContext
 from pydantic import AnyUrl
-from pytest_simcore.helpers.logging_ext import test_logger
+from pytest_simcore.helpers.logging_tools import test_logger
 from tenacity import Retrying
 from tenacity.retry import retry_if_exception_type
 from tenacity.stop import stop_after_delay
