+++ tests/e2e-playwright/tests/sim4life/test_sim4life.py
@@ -13,7 +13,7 @@ from typing import Final
 
 from playwright.sync_api import APIRequestContext, Page
 from pydantic import AnyUrl
-from pytest_simcore.helpers.logging_ext import test_logger
+from pytest_simcore.helpers.logging_tools import test_logger
 from pytest_simcore.helpers.playwright import on_web_socket_default_handler
 from tenacity import Retrying
 from tenacity.retry import retry_if_exception_type
