+++ tests/e2e-playwright/tests/sleepers/test_sleepers.py
@@ -17,7 +17,11 @@ from typing import Any, Final
 from packaging.version import Version
 from packaging.version import parse as parse_version
 from playwright.sync_api import Page, WebSocket
-from pytest_simcore.helpers.logging_ext import ContextMessages, log_context, test_logger
+from pytest_simcore.helpers.logging_tools import (
+    ContextMessages,
+    log_context,
+    test_logger,
+)
 from pytest_simcore.helpers.playwright import (
     MINUTE,
     RunningState,
